:grin:  Added eagleUp folder, modded jack.lbr
I'm testing 3D view of the PCB with components on it because it's cool
and because it'll let me see if the components would overlap in real
life. I'm also modding the jack.lbr file to show little squares around
little parts on silkscreen, to triple-check packages sizes, and to match
packages names with eagleUp model names. I'll eventually add the final
SketchUp file of it.